EKF: Disable relevant status checks
This is a temporary solution to enable local position messages from
ArduPilot. The status checks are there to see if we have enough
information to provide even more accurate local positioning using GPS,
flow sensor or wind sensor, and also to disable the EKF when the rover
is standing still or we have no GPS. However, the raw data from
gyroscope is already quite accurate after it settles during moving. Thus
it is helpful to actually have the message in this case.